refactor: CategoryClient component for categories page
